modules: supporting parameters and maybe an established concept
the idea:
a module is a chicken-module. it has a single .scm file with it. it
has:
- name
- path (default ./name)
- file (default name.scm)

so a module 'bind has its source in ./bind/bind.scm
a module (coops file: coops-module) has its source in
./coops/coops-module

see module definition. this should be all we need to compile scm to .c
files and generate android .mk files for them and their import files